Backup of former duck branch of Fairy-Stockfish, with added duck-custom endgames and piece values
This is a backup of the duck branch of Fairy-Stockfish by Fabian Fichter, from 2023-Dec-31. Custom endgames and piece values for duck chess added by Michael Taktikos. This additions improve duck chess, but weaken other variants, so use this repo for duck chess only! Also further updates from FSF main branch are incompatible with this. Therefore it is published as non-updateable, standalone repository.